Change type of price from OtoMoto

+++ OtoMotoWebFlowSync/Model/OtoMoto/Advert.cs
@@ -78,7 +78,7 @@ public class Price
     [JsonPropertyName("0")]
     public string? Zero { get; set; }
     [JsonPropertyName("1")]
-    public int? Amount { get; set; }
+    public double? Amount { get; set; }
     public string? Currency { get; set; }
     public string? GrossNet { get; set; }
 }
